♻️ refactor: reorganize migration scripts directory structure
