+++ scripts/clerk-to-betterauth/_internal/config.ts
