+++ scripts/clerk-to-betterauth/_internal/env.ts
