+++ scripts/clerk-to-betterauth/_internal/load-data-from-files.ts
