+++ scripts/clerk-to-betterauth/_internal/types.ts
