+++ scripts/clerk-to-betterauth/_internal/utils.ts
